Refactor histogram
